modified:   dataset/__init__.py 	modified:   dataset/dataset.py 	modified:   main_single_gpu.py 	modified:   model/model.py 	modified:   train_eval/__init__.py 	modified:   train_eval/train_eval_cls.py 	modified:   train_eval/train_eval_recover_rgb.py 	modified:   utils/__init__.py

+++ dataset/dataset.py
@@ -81,6 +81,6 @@ class Recover_rgb_dataset(data.Dataset):
     def collate_fn(batch: list) -> tuple:
         data, rgb = list(zip(*batch))
         data = torch.stack(data, dtype=torch.float32)
-        rgb = torch.stack(rgb, dtype=torch.float32)
+        rgb = torch.stack(rgb, dtype=torch.float32).long()
         return data, rgb
